first commit
promotion menu is not finished.

promblem: board and pieces use collision box to detect where player
want to move pieces, this is not good because players could use that
collision to jump in places where they shouldn't be able to.

but to avoid that, need to develop custom collision detection with
raycasting, so that my custom rays will collide with board and pieces
but not the player.